pyproject.toml
[project]
name = "langgraph-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "factory-boy>=3.3.3",
    "faker>=37.5.3",
    "ipykernel>=6.30.1",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.3",
    "langgraph-cli[inmem]>=0.3.6",
    "langgraph-prebuilt>=0.6.3",
    "langgraph-sdk>=0.2.0",
    "langsmith>=0.4.13",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
